Update dice calculations to allow finding max roll
Changed the ForceTrait from SR5E to do addition and subtraction
correctly, left to right, instead of doing all of the addition before
all of the subtraction. Similarly fixed multiplication and division.

Stillfleet uses the max for a given attribute, which may be something
like d8+1, so it needs to be able to calculate that.