Fix creating characters for modules without character classes
Tests are failing because the generic character factory uses the list of
supported systems to create characters, but not all modules have a class
for their character sheet.

Also some random cleanups, switching from using backslashes to imports
and expanding the list of systems that use "handle" instead of "name" in
the character factory.