Downgrade nwidart/laravel-modules
It seems to have stopped working on Commlink with 11.1.0 with a
ModuleNotFoundException. I've created a stuf project to try to reproduce
it but everything seems to work okay. I'm wondering if the way I
upgraded from Laravel 10 to 11 broke some convention. I'll try and
revisit this after fixing some of the things that Laravel Shift brought
up.